refactor(schedules): move the sources under Appwrite\\Schedule\\Source
Source/Database.php holds what reading the schedules collection takes, and
Source/{Functions,Executions,Messages}.php say what kind. Call sites read
new Source\\Executions(...), which is the library layout they mirror:
Utopia\\Schedule\\Source and Store\\Redis are named for what backs them, not
for the interface they satisfy.

The base cannot import Utopia\\Database\\Database while being called Database
itself, so the platform connection is written out in the three signatures
that take it rather than aliased to a word nothing else here uses.
