+++ src/Appwrite/Platform/Tasks/ScheduleFunctions.php
@@ -4,7 +4,7 @@ namespace Appwrite\Platform\Tasks;
 
 use Appwrite\Event\Message\Func as FunctionMessage;
 use Appwrite\Event\Publisher\Func as FunctionPublisher;
-use Appwrite\Schedule\FunctionSchedule;
+use Appwrite\Schedule\Source;
 use Utopia\Database\Database;
 use Utopia\Database\DateTime;
 use Utopia\Database\Document;
@@ -36,7 +36,7 @@ class ScheduleFunctions extends Action
 
     public function action(FunctionPublisher $publisherForFunctions, callable $getIsResourceBlocked, Database $dbForPlatform, callable $getProjectDB, Telemetry $telemetry): void
     {
-        $source = new FunctionSchedule(
+        $source = new Source\Functions(
             $dbForPlatform,
             $getProjectDB,
             $getIsResourceBlocked,
